requirements.txt
discord.py==0.16.12
textgenrnn==1.5
boto3==1.9.106
sqlalchemy==1.3.0
psycopg2==2.7.7
pandas==0.24.1
matplotlib==3.0.3
wordcloud==1.5.0
s3fs==0.2.0
multidict==4.5.2